Homework 5

+++ src/homework/05_functions/func.h
@@ -1,6 +1,6 @@
 #include <string>
 //add function(s) prototype here
-int get_gc_content(const std:: string& dna);
-//void get_dna_complement(string dna);
+double get_gc_content(const std:: string& dna);
+//std :: string get_dna_complement(string dna);
 // void reverse_string(string dna);
